requirements.txt
asposestorage==1.0.2
diffusers==0.16.1
h5py==3.10.0
image_processor==0.1.1
ipython==8.20.0
ipython==8.12.3
lpips==0.1.4
neptune==1.10.4
numpy==1.25.2
opencv_python==4.9.0.80
packaging==24.1
pandas==2.2.2
Pillow==10.1.0
Pillow==10.3.0
pipeline_utils==0.0.2
pycocotools==2.0.7
safety_checker==0.5
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.27.4
